fix(db): fix point calculation bug
Drizzle returned a string when using count
(drizzle-team/drizzle-orm/issues/999#issuecomment-1669266023).

+++ src/lib/server/cron.ts
@@ -59,7 +59,7 @@ export const updateChallengePoints = async () => {
   const challenges = await db
     .select({
       id: schema.challenges.id,
-      solvesCount: sql<number>`count(${schema.solves.id})`,
+      solvesCount: sql<number>`count(${schema.solves.id})::int`,
     })
     .from(schema.challenges)
     .leftJoin(
